test: 전체 Controller를 사용하는 test를 한 클래스로 통합
