--- src/test/java/racingcar/controller/ExceptionTest.java
@@ -1,49 +1,0 @@
-package racingcar.controller;
-
-import io.restassured.RestAssured;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import racingcar.dto.RacingGameInputDto;
-
-import static org.hamcrest.Matchers.containsString;
-
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class ExceptionTest {
-    @LocalServerPort
-    int port;
-
-    @BeforeEach
-    void setup() {
-        RestAssured.port = port;
-    }
-
-    @Test
-    void exceptionHandler1() {
-        final RacingGameInputDto racingGameInputDto = new RacingGameInputDto("포비,  ,구구", 10);
-
-        RestAssured.given().log().all()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(racingGameInputDto)
-                .when().post("/plays")
-                .then().log().all()
-                .statusCode(HttpStatus.BAD_REQUEST.value())
-                .body(containsString("[ERROR]"));
-    }
-
-    @Test
-    void exceptionHandler2() {
-        final RacingGameInputDto racingGameInputDto = new RacingGameInputDto("포비,브라운,구구", 101);
-
-        RestAssured.given().log().all()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(racingGameInputDto)
-                .when().post("/plays")
-                .then().log().all()
-                .statusCode(HttpStatus.BAD_REQUEST.value())
-                .body(containsString("[ERROR]"));
-    }
-}
