+++ src/test/java/racingcar/controller/RacingGameControllerTest.java
@@ -3,16 +3,21 @@ package racingcar.controller;
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
+import org.springframework.test.context.jdbc.Sql;
 import racingcar.dto.RacingGameInputDto;
 
+import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.core.Is.is;
 import static org.hamcrest.core.IsNull.notNullValue;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+/*@JdbcTest
+@Sql({"test_data.sql"})*/
 public class RacingGameControllerTest {
     @LocalServerPort
     int port;
@@ -34,5 +39,31 @@ public class RacingGameControllerTest {
                 .statusCode(HttpStatus.OK.value())
                 .body("winners", notNullValue())
                 .body("racingCars.size()", is(3));
+    }
+
+    @Test
+    void exceptionHandler1() {
+        final RacingGameInputDto racingGameInputDto = new RacingGameInputDto("포비,  ,구구", 10);
+
+        RestAssured.given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(racingGameInputDto)
+                .when().post("/plays")
+                .then().log().all()
+                .statusCode(HttpStatus.BAD_REQUEST.value())
+                .body(containsString("[ERROR]"));
+    }
+
+    @Test
+    void exceptionHandler2() {
+        final RacingGameInputDto racingGameInputDto = new RacingGameInputDto("포비,브라운,구구", 101);
+
+        RestAssured.given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(racingGameInputDto)
+                .when().post("/plays")
+                .then().log().all()
+                .statusCode(HttpStatus.BAD_REQUEST.value())
+                .body(containsString("[ERROR]"));
     }
 }
